requirements.txt
numpy
opencv-python
protobuf
easydict
pytorch
pillow
matplotlib
# wxpython
shapely
tqdm
open3d